UI Hotfix

+++ Assets/02_Scripts/UI/UIBattle/UIBossHpsManager.cs
@@ -43,7 +43,7 @@ public class UIBossHpsManager : MonoBehaviour
     {
         textMeshPro.text = (curHp + "/" + maxHp); // 텍스트 업데이트 "현재체력/최대체력"
         Vector2 newCurHP = images[0].rectTransform.sizeDelta;
-        float hpRatio = curHp / maxHp;
+        float hpRatio = (float)curHp / (float)maxHp;
         newCurHP.x = hpRatio * newCurHP.x;  // X 값만 변경
         images[images.Count - 1].rectTransform.sizeDelta = newCurHP;  // 변경된 sizeDelta 적용
     }//UI 표시 함수
